+++ tests/e2e/common/mod.rs
@@ -1,2 +1,3 @@
 pub mod containers;
 pub mod logs;
+pub mod setup;
